Introduce UI enhancements, caching improvements, emails, and CI workflow updates (#54)
* cleanup darkmode.js file and add darkmode toggle to navbar

* Update prettier rc rules and and run linter, php/json 4 spaces, js/vue/yml 2 spaces.

* update service providers to be separated and add Laravel Debugbar for ONLY local environments when enabled and debug enabled

* add user call caching with busting observers

* Added Personalisation Model front end caching with observer to bust on events. Cleanup service providers.

* Added  var to disable profile disable/delete for demo purposes.

* Apply fixes from StyleCI

* Add GHA actions for Laravel, node, eslint, dependency reviews, and code scanning for vulnerabilities.

* cleanup 40x/50x template, cleanup minor app.js, add clean and optimize npm commands

* file pond uploader bug fixes

* cleanup app.js

* update Inertia::share to get data in one place, AppPersonalisationServiceProvider, update the uploader

* Apply fixes from StyleCI

* fixed missing jobs table for database queue, added goodbye email for users deleting their account using events and listeners

* cleanup

* Apply fixes from StyleCI

* lint

* Apply fixes from StyleCI

* update clean command

* Add welcome email

* lint

* fix email blades

---------

+++ app/Http/Controllers/UserAccountController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Events\UserDeleted;
 use App\Models\User;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
@@ -30,6 +31,8 @@ class UserAccountController extends Controller
             'twoFactorEnabled'  => Route::has('two-factor.enable'),
             'passwordEnabled'   => Route::has('user-password.update'),
             'sessions'          => $this->getUserSessionsData($user, $request->session()->getId()),
+            'deactivateEnabled' => config('guacpanel.user.account.deactivate_enabled'),
+            'deleteEnabled'     => config('guacpanel.user.account.delete_enabled'),
         ];
 
         return Inertia::render('UserAccount/IndexPage', $data);
@@ -147,6 +150,10 @@ class UserAccountController extends Controller
 
     public function deactivateAccount()
     {
+        if (!config('guacpanel.user.account.deactivate_enabled')) {
+            return redirect()->back()->with('error', 'Feature disabled in the .env file');
+        }
+
         $user = Auth::user();
         $user->update(['disable_account' => true]);
 
@@ -155,12 +162,19 @@ class UserAccountController extends Controller
         $request->session()->invalidate();
         $request->session()->regenerateToken();
 
-        return redirect()->route('login')->with('info', 'Account has been deactivated successfully.');
+        return redirect()->route('home')->with('success', 'Account has been deactivated successfully.');
     }
 
     public function deleteAccount()
     {
+        if (!config('guacpanel.user.account.delete_enabled')) {
+            return redirect()->back()->with('error', 'Feature disabled in the .env file');
+        }
+
         $user = Auth::user();
+
+        UserDeleted::dispatch($user);
+
         $user->delete();
 
         Auth::logout();
@@ -168,6 +182,6 @@ class UserAccountController extends Controller
         $request->session()->invalidate();
         $request->session()->regenerateToken();
 
-        return redirect()->route('login')->with('info', 'Account has been deleted successfully.');
+        return redirect()->route('home')->with('success', 'Account has been deleted successfully.');
     }
 }
